카드 컴포넌트에 Link를 통한 라우팅 기능 구현 (#7)
* feat: 카드 컴포넌트의 Link를 사용하여 라우팅할 수 있는 기능 구현

* refactor: cardProps를 관리하는 파일 추가

* refactor: 마법의 고민해결 책 카드 컴포넌트 수정

내부 문구 수정
테스트 썸네일에서 실제로 사용할 썸네일로 변경

* refactor: card의 썸네일 alt 부분을 '타이틀 thumbnail' 형식으로 변경

* refactor: assets 내부의 이미지들을 디렉토리 안에 구조화

* refactor: 파일 이미지 오타 수정

* refactor: home에서 관리하던 카드 컴포넌트의 데이터를 data.ts로 이동
