+++ src/components/Card/Card.tsx
@@ -1,20 +1,16 @@
 import styles from './Card.module.css';
+import { Link } from 'react-router-dom';
+import CardProps from '@/types/card.ts';
 
-interface CardProps {
-  thumbnail: string;
-  title: string;
-  description: string;
-}
-
-function Card({ thumbnail, title, description }: Readonly<CardProps>) {
+function Card({ thumbnail, title, description, link }: Readonly<CardProps>) {
   return (
-    <div className={styles.cardContainer}>
-      <img className={styles.image} src={thumbnail} alt="thumbnail" />
+    <Link className={styles.cardContainer} to={link}>
+      <img className={styles.image} src={thumbnail} alt={`${title} thumbnail`} />
       <div className={styles.textContainer}>
         <h3>{title}</h3>
         <p className={styles.description}>{description}</p>
       </div>
-    </div>
+    </Link>
   );
 }
